deletion should not actually remove wc file if it is locally modified, just remove admin entries

+++ javasvn/src/org/tmatesoft/svn/core/internal/ws/fs/FSDirEntry.java
@@ -641,7 +641,11 @@ public class FSDirEntry extends FSEntry implements ISVNDirectoryEntry {
                 doDeleteFiles(child);
             }
         }
+        boolean keepWC = !entry.isDirectory() && entry.asFile().isContentsModified();
         getAdminArea().deleteArea(entry);
+        if (keepWC) {
+            return;
+        }
         File file = getRootEntry().getWorkingCopyFile(entry);
         if (file != null && file.exists()) {
             file.delete();
